01 - Reading List: Refactored code for modal feature.
