+++ pruebas/01-reading-list/iivvaannxx/src/features/modal/components/modal.svelte
@@ -1,7 +1,7 @@
 <script lang='ts' context='module'>
 
-  import { isOpen } from './store'
-  import { CloseIcon } from '$assets/icons';
+  import { isOpen, title, close } from '../store'
+  import Header from './header.svelte';
 
 </script>
 
@@ -22,50 +22,29 @@
 
   $: show = $isOpen
 
-  /** @brief Hides the modal window. */
-  function hide () {
-
-    show = false
-  }
-
 </script>
 
 <!-- svelte-ignore a11y-no-noninteractive-element-interactions -->
 <dialog class='flex-col fixed w-1/3 h-1/2 inset-0 rounded-xl border-none backdrop:bg-black backdrop:bg-opacity-30 open:animate-zoom-in open:backdrop:animate-fade-in'
 
   bind:this={dialog}
-  on:close={hide}
-  on:keypress={(e) => { if (e.key === 'Escape') hide() }}
+  on:close={close}
+  on:keypress={(e) => { if (e.key === 'Escape') close() }}
 
-  on:click|self={hide}
+  on:click|self={close}
   on:click|stopPropagation
 >
 
   <div class='flex flex-col h-full w-full'>
 
-    <header class='h-16 flex flex-row p-4 justify-between'>
-
-      <div class='h-full flex items-center ml-4'>
-        <slot name='header' />
-      </div>
-
-      <div class='h-full flex items-center mr-4'>
-
-        <button tabindex="0" on:click={hide}>
-          <CloseIcon />
-        </button>
-
-      </div>
-
-    </header>
+    <Header title={$title} onShouldClose={close} />
 
     <hr />
 
     <div class='flex-1 p-4'>
-      <slot name='content'/>
+      <slot />
     </div>
 
   </div>
-
 
 </dialog>
