+++ pruebas/01-reading-list/iivvaannxx/src/features/modal/store.ts
@@ -1,12 +1,24 @@
 import { atom } from 'nanostores'
 
-export const isOpen = atom(false)
+/** @brief Whether the modal dialog is opened or not. */
+export const isOpen = atom(true)
 
+/** @brief The title of the modal dialog. */
+export const title = atom('No Title')
+
+/** @brief Sets a new title for the modal dialog. */
+export const setModalTitle = (newTitle: string) => {
+
+  title.set(newTitle)
+}
+
+/** @brief Opens the modal dialog. */
 export function open () {
 
   isOpen.set(true)
 }
 
+/** @brief Closes the modal dialog. */
 export function close () {
 
   isOpen.set(false)
